requirements.txt
srt
tmdbsimple
opensubtitlescom
python-ffmpeg
